Issue #1: View and update membership info: Logging + Misc
Moved the serializers to a separate file.

Renamed urls-api and urls-frontend to urls_api and urls_frontend respectively (to allow importing).
Split views.py into views_api.py and views_frontend.py respectively.

Updated the Member model:
- Changed the regex on TU/e card numbers (must now consist of exactly 7 numbers)
- Changed the regex on external card numbers (must now consist of exactly 7 numbers, followed by a hyphen, followed by exactly 3 numbers. I.e. it now also includes the external number)
- Added an external card cluster field (thtat is present on some external cards, but not all of them)
- Added state/province field (that may be needed for members living abroad)
- Added country field (that may be needed for members living abroad)
- updated postal code regex to accept postal codes consisting of letters, numbers, and spaces. As such, it no longer accepts only Dutch postal codes
- Updated the member_since field to allow it to be re-defined rather than read-only.
- Added a last-updated-by field to signify the user that last updated the member.

Every time a change to a member is made (incl. member creation), this is logged in a MemberLog.
Updated Memberlog to now include the user that updated the member, the member that was updated, and the date/time at which this update was performed.

Updated fields are captured in a MemberLogField, consisting of a MemberLog-id, the field name, the old value of the field, and the new value of the field.