Issue #4: Unit Tests: Set up Coverage.py
Set up Coverage.py to test code coverage. It's also possible to output it in a nice .html format.
The Coverage.py configuration can be edited in the .coveragerc file.

Split the dependencies into development dependencies and production dependencies, and added a disclaimer comment in both files to indicate this.

Updated the README to reflect the changes above.