Issue #1: View and update membership info: Testcases for deletion logic + misc
Added testcases for deletion logic of members. That is, it tests whether:
- a member can NOT be deleted if it was NOT marked for deletion
- a member can NOT be deleted if it was marked for deletion by the user who attempts to delete the member
- a member can be deleted if it was marked for deletion by a user other than the one who attempts to delete the member
- a member's information cannot be updated if it is marked for deletion (except for its marked for deletion status)

Fixed a bug where a member could be deleted by anyone if it was marked for deletion and last_updated_by = None.

Changed the memberlog testcases to check for 200 OK responses, rather than only checking if a 500 Internal Server response was not found.