Issue #4: Unit Tests: Requirement-files + Misc
Moved the requirement-files into a separate folder, and made requirements/dev.txt and requirements/prod.txt automatically include the common requirements (such as Django).

Updated the README to reflect this change.

Ensured that a /coverage-folder is created if it does not exist upon running the application. Without it, coverage.py will crash when attempting to run the testcases.