Issue #5: Login, Logout, Register: Logging in and out
Users with an existing account can now log in and out.
Added a page that users can only visit once they have logged in. If they attempt to access the page without having logged in, they are redirected to the login page.

If a user logs out, they are redirected to a page displaying that they have successfully logged out. If they access this page while logged in, they will be redirected to the logout page first.