Issue #5: Login, Logout, Register: Registering + Misc
Users can now register an account. When registering, a (unique) email, a (unique) username, an (optional) nickname, and a password field are shown.
An additional password-field validates whether the two passwords match (and thus no typo was made).

Added a decorator to util.py that surpresses warnings when running testcases (which can be useful if such warnings are triggered on purpose)