Issue #26: Members view and update info: Misc
Added student number and educational institution to the membership_file model. Removed postal_code from that same model (as this can be inferred from the address).

Cleaned up unused URL patterns.

Added a membership_required decorator that redirects to a specific URL if the requesting user is not a member.