Issue #26: Members view and update info: Testcases
Added most testcases for all functionality added in this branch.
Currently only missing POST-request tests.

Split the test-files of the membership_file app into two separate files. Ensured that files containing the word "tests" are excluded from the coverage report.

Fixed a bug where the submit button was dashed on the Edge browser. None of the input fields are now dashed though. There seems to be a bug in the edge-browser as manually editing the css in the developer panel did work properly.

Changed get_last_updated_name to display_last_updated_name to be in line with the other member-methods.