Issue #35: General Frontend: Updated view membership page
Update the view membership data page such that it...
- ...is formatted nicely
- ...contains member initials as well

A member can no longer edit its information if it is marked for deletion.
A shortcut to the membership information was added to the account-dropdown, but only if the logged in user is a member.

Added help text to the phone number and email fields of the membership file.