Issue #38: Final pre-release changes: Achievement update
Added a priority to the Categories so that they can be manually sorted if desired.
Added an option to hide achievements, meaning that they can only be accessed inside the admin panel.
The achievementserializer now properly returns the number of claimants, even if a user is not logged in.

Only members can view achievement claimants.