issue #48: iCal feed: Front-end + bugfixes
Ensured that the events are correclty loaded in FullCalendar. The FullCalendar frontend now (always seems to) match how events are displayed in Google Calendar.

Fixed a bug where an EXDATE's time did not match the event's time, resulting in the event's occurence not being excluded on the signified day. Note: This is actually a bug in the recurrence library.

Ensured that it is possible to pass a value for an item's RECURRENCE-ID field.

Removed unneeded code, debug code, and comments.