Issue #50: activity subscriptions: Model + helper methods + frontend
Added a Participant and Slot-model. Each activity can have a number of slots that users can register for.

The front-end indicates whether an activity requires registration, whether you are subscribed, and how many participants are currently registered and can at most be registered.

Activity settings can limit the registration period, the maximum number of slots, and the maximum number of participants.

Added helper methods for obtaining an activity's participants and slots. This takes the activity-settings into account.