Issue #50: activity subscriptions: register page + images + mobile
Improved mobile formatting across the entire application. Fixed a problem on the achievements page where their icons were not shown due to this new formatting.

Added a page on which all slots for a certain activity occurence are listed.

Slots (and activties) now have an image, which is shown on the register page. A default image is used if none is provided.

All .sqlite3 files are now ignored. This makes debugging easier.