Issue #60: browser compatibility: Mobile navbar
The navigation bar now has a proper height and properly expands on mobile devices. The footer remains at the bottom of the page even if there is not enough content. (Fixed by changing the height 100% property of the body to min-height 100%)

Added a left and right margin to the page's main content, which only really makes a difference for mobile devices layout.

Reduced the size of the images on mobile devices.