Deregister Members + bugfix
Fixed a bug where a member's user could not be set in the admin UI.

Removed start and end date fields from the admin's activityslot UI as those are not supported.

Added an option to "deregister" a member. This allows keeping (contact) information while showing the relevant user that it is no longer a member.