Override "-----" and "unknown" in ActivityMomentForm
The null-options for Select- and NullBooleanSelect-fields are now overriden in the ActivityMoment form. Fields are no longer modified to enable the correct placeholder text.

Replaced the exclude field of the ActivityMomentForm's Meta-class by a "fields" field, such that it is possible to specificy the order in which the fields are rendered in the form.

Regenerated the relevant migration (which was only introduced in this branch).