EXDATE and RDATE time match start time of activity
RDATEs no longer show up occurring at 00:00h in the ical feed.
An EXDATE that previously did not work does work now.

Modifying the time of the rdates/exdates is done in the same manner for the FullCalendar as the ical feed.

Should still add more testcases (specifically for the CET to CEST switch)