MarkdownImage uploads for new models + refactor
Refactored to reduce code duplication and to make it easier to add more models supporting MarkdownImages.

When saving a new model instance, all MarkdownImages uploaded for that particular model by the current user, but without any assigned instance, are assigned to the newly created instance. This reduces the amount of "orphan" MarkdownImages that can exist.