Addressed Review Comments
Renamed several classes. Revisited the MembershipRequiredMixin logic. It now automatically includes the LoginRequiredMixin, making behave identically to the @-membership_required decorator.

Removed the check for the exclude field in the MemberRoomForm's Metaclass, as it would not make sense to use this form without wanting to use the only field provided by it