Refactored RequestUserForm and friends
Created a new Mixin for ModelAdmins that passes the requesting user to that ModelAdmin's form as a kwarg, as this was currently being done three times separately. There does not seem to be a straightforward way to pass form kwargs to Forms of ModelAdmins.

Created a RequestUserToFormViewMixin that does the same for CBVs, and added it to the EditActivityMomentView where this was still missing.

Renamed RequestUserFormMixin to UserFormMixin, as there's no requirement for the passed user to be the requesting user.

Made UpdatingUserFormMixin inherit from UserFormMixin. Also fixed a bug where its updating_user_field_name was already set at form initialisation, rather than upon Form saving, meaning that the old value for this field was never shown.

Moved them all to utils.forms