Alt start time notification
Adds extra text on occurrence pages (and the activitymoment edit page) if that occurrence has a different start date/time than it would normally have.

If the occurrence takes place on a different day than normal, the full original date is shown. If it occurs on the same date, only the time difference is shown instead.