Bugfix: Recurrences at 00:00-02:00
Weekly recurrences that start between 00:00-02:00 (CEST) now have recurrences on the correct day instead of a day later.

dtstart passed into recurrences.between was in UTC (not working), but now in the local time (working). This is probably a bug in django_recurrences. Should still add a testcase for future reference.