Bugfix Martor CSS + Markdown Admin + Achievement Admin queries
Changed the __str__ method of Claimant objects to no longer fetch both the achievement and user (+ member) from the database.
This was especially bad for achievements with many claimants.

Fixed a bug where Martor's CSS revealed an input that should remain hidden in the admin panel (See #139 )

Allowed image uploads for Committee Instructions and Roleplay Systems.