Register Bugfixes + tests
Fixed a bug where it was not possible to create slots when the relevant activitymoment was not yet created in the database.

Added testcases for registering for activities (slots and slotless) when their activitymoments are not yet in the database.

Fixed an ancient race condition that could create two slots when two users register for slotless activities at the same time, instead of only one.