Upgrade packages
Django 2.2 has reached its end of life, and many packages don't officially support it anymore.
I'm taking this as an opportunity to upgrade all of them, though we should really do this more often in the future, especially for security fixes.

The current failing testcase is a direct result of upgrading to Django 3.2, and not because of a package upgrade.
The changes in membership_file/admin.py prevent more errors, also caused by the bump in Django version.