Disable admin dark theme + Martor CSS/JS
Django introduced a Dark mode for the admin panel, but various of our widgets (added by libraries) don't support it. The dark theme is disabled to "fix" this.

Furthermore, jQuery is no longer automatically loaded in the admin panel.

Martor proves to be a pain once again with its jQuery and Bootstrap versions. It uses newer versions of jQuery ánd Bootstrap (bootstrap 5) than Squire. Furthermore, Bootstrap5 no longer uses jQuery in the first place. Ideally we unify these in the future.

Modified the way that Martor and the admin panel's CSS interact. Martor relies on bootstrap, which previously broke the admin CSS.
Rather than reverting it, we now only apply the Bootstrap CSS to Martor's widget.
This does require us to compile some SCSS, for which I have not created a development setup. Moving to SCSS is a stretch goal (see #196 ), but not for this PR.