config + user preference
Moved the mailcow config to a separate json file, and added this to the gitignore. Included a sample config to show the formatting.

Added a page that allows users to set their preferences for each of the member aliases. Aliases that do not allow opt-outs are still shown, though their settings are disabled.