Committee Aliases status + split template
Allow updating and viewing the status of committee aliases. If an alias is not managed by Squire, it now instead lists the emails present in Mailcow.

Pre-existing mailboxes break the system, as Squire still attempts to create an alias for those.

Split off a separate alias-list template from the status template.