Link account after login
Allow attaching membership to an already existing Squire account by re-entering login credentials. Doing so also syncs the member's and user's real name and email address.

Added a LoginView to allow subclassing. When given an initial username, the form also cannot be edited.
Added a LinkedLoginView that displays a login form that indicates that the attached form will link a Squire account to some other piece of data, such as membership data.

Refactored the logoutSuccess page into a class-based view.