Added proper fail page
Fail page is reached when the token is invalid, or when an account link would otherwise be invalid. E.g. the member already has a linked user, the logged in user already as an associated member, a user is logged in and tries to register an account.

Also cleaned up debugging statements.