Update AssociationGroupsTabAccessForm to use ModelForm
Context:
The UUPS new member group, which [person] is part of, poses one new situation that we haven't had before. What if someone can view the page, but does not have rights to edit everything.
The Django Admin is very simple in that aspect. We render everything as is, but on the linked Form (normally a ModelForm) we set every field in the Forms' readonly_fields property instead.
During the render process, the admin notices the readonly fields and try to render them as such - by the way, the 'default' values that we as superusers can see are rendered through a different pipeline as in that case it's a special field instances that knows it is default and thus renders the disabled state - anyhow, read only fields in the admin are rendered through the AdminReadonlyField class and that is where _meta is called. But... it's not a ModelForm so fails

I'm not in a position to test the code today, but I think if you adjust the AssociationGroupsTabAccessForm to the following code it should be fixed.

Co-authored-by: Lena Wildervanck <lena@wildervanck.eu>